lectute 2 completed

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 //@desc Get goals
 //@route GET /api/goals
 //@access Private
 
 const getGoals = asyncHandler(async (req,res)=>{
-  const goals = await Goal.find()
+  const goals = await Goal.find({user: req.user.id})
   res.status(200).json(goals)
 })
 
@@ -21,7 +22,8 @@ const setGoals = asyncHandler(async (req,res)=>{
     throw new Error('please add a text field')
   }
   const goal  = await Goal.create({
-    text:req.body.text
+    text:req.body.text,
+    user: req.user.id,
   })
 
   res.status(200).json({goal})
@@ -37,6 +39,24 @@ const updateGoals = asyncHandler(async (req,res)=>{
   res.status(400)
   throw new Error('Goal not found')
  }
+ const user  = await User.findById(req.user,id)
+
+//  check for user
+if(!user){
+  res.status(401)
+  throw new Error('User not found')
+
+}
+// make sure  the logged in user matches the goal user
+if(goal.user.toString()!=user.id){
+  res.status(401)
+  throw new Error('Not authorized to update this goal')
+
+
+}
+
+
+
  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body,{new:true})
   res.status(200).json({updatedGoal})
 }
@@ -51,6 +71,24 @@ const deleteGoals = asyncHandler(async (req,res)=>{
   res.status(400)
   throw new Error('Goal not found')
  }
+
+ const user  = await User.findById(req.user,id)
+
+//  check for user
+if(!user){
+  res.status(401)
+  throw new Error('User not found')
+
+}
+// make sure  the logged in user matches the goal user
+if(goal.user.toString()!=user.id){
+  res.status(401)
+  throw new Error('Not authorized to delete this goal')
+
+
+}
+
+
 
  await Goal.findByIdAndDelete(req.params.id);
